backend-requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
httpx==0.25.0
aiofiles==23.2.1
networkx==3.1
numpy==1.26.0
pandas==2.1.1
transformers==4.35.0
optimum==1.13.1
optimum-conda==0.1.1
onnxruntime==1.16.0
onnxruntime-gpu==1.16.0
tiktoken==0.5.1
safetensors==0.4.0
accelerate==0.23.0
bitsandbytes==0.41.1
sentencepiece==0.1.99
protobuf==4.24.4
docker==6.1.3
python-dotenv==1.0.0
tenacity==8.2.3
aiohttp==3.8.6
backoff==2.2.1
python-dateutil==2.8.2
Jinja2==3.1.2
Pillow==10.0.1
matplotlib==3.8.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
huggingface-hub==0.17.3
